requirements.txt
streamlit==1.37.1
torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1
Pillow==10.4.0
numpy==1.26.4
